+++ docs/src/SUMMARY.md
@@ -7,8 +7,8 @@
   - [Basic usage](./getting-started/basics.md)
   - [Wallets](./getting-started/wallets.md)
   - [Generating type-safe Rust bindings](./getting-started/type-safe-bindings.md)
+- [Testing](./testing/index.md)
 - [`fuels-rs` Rust Workspaces](./workspaces/index.md)
   - [`fuels-abigen-macro`](./workspaces/fuels-abigen-macro.md)
   - [`fuels-abi-cli`](./workspaces/fuels-abi-cli.md)
-- [Testing](./testing/index.md)
