+++ docs/src/testing/run-script.md
@@ -1,11 +1,11 @@
 # `run_script` 
 
-`run_script` is helper function for testing simple sway code and reducing boilerplate code in tests.
+`run_script` is helper function for testing simple Sway scripts and reducing boilerplate code related to setting up contracts and deployment.
 As a parameter the function receives a `&str`, that is, the path that leads to the generated `.bin` file.
 
 ## Usage
 
-A simple example of testing using `run_script`
+You can use it this way:
 
 ````rust
 #[tokio::test]
