+++ packages/fuels-abigen-macro/tests/harness.rs
@@ -1436,21 +1436,3 @@ async fn test_wallet_balance_api() {
     assert!(balances.contains_key(&expected_key));
     assert_eq!(*balances.get(&expected_key).unwrap(), 231)
 }
-
-#[tokio::test]
-async fn test_require_sway() {
-    let path_to_bin = "tests/test_projects/require/out/debug/require.bin";
-    let return_val = run_script(path_to_bin).await;
-    assert_eq!(1, return_val[0].val().unwrap());
-}
-
-#[tokio::test]
-async fn test_logging_sway() {
-    let path_to_bin = "tests/test_projects/logging/out/debug/logging.bin";
-    let return_val = run_script(path_to_bin).await;
-
-    let correct_hex =
-        hex::decode("ef86afa9696cf0dc6385e2c407a6e159a1103cefb7e2ae0636fb33d3cb2a9e4a");
-
-    assert_eq!(correct_hex.unwrap(), return_val[0].data().unwrap());
-}
