--- packages/fuels-abigen-macro/tests/test_projects/require/.gitignore
@@ -1,3 +1,0 @@
-out
-target
-Forc.lock
