--- packages/fuels-abigen-macro/tests/test_projects/require/Forc.lock
@@ -1,11 +1,0 @@
-[[package]]
-name = 'core'
-dependencies = []
-
-[[package]]
-name = 'require'
-dependencies = ['std']
-
-[[package]]
-name = 'std'
-dependencies = ['core']
