--- packages/fuels-abigen-macro/tests/test_projects/require/Forc.toml
@@ -1,7 +1,0 @@
-[project]
-authors = ["Fuel Labs <contact@fuel.sh>"]
-entry = "main.sw"
-license = "Apache-2.0"
-name = "require"
-
-[dependencies]
