+++ packages/fuels-test-helpers/Cargo.toml
@@ -10,9 +10,9 @@ description = "Fuel Rust SDK test helpers."
 
 [dependencies]
 fuel-core = { version = "0.8", default-features = false }
-fuel-crypto = "0.5"
 fuel-gql-client = { version = "0.8", default-features = false }
 fuel-tx = "0.12"
+hex = { version = "0.4.3", default-features = false, features = ["std"] }
 fuel-types = { version = "0.5", default-features = false }
 fuels-contract = { version = "0.14.1", path = "../fuels-contract" }
 fuels-signers = { version = "0.14.1", path = "../fuels-signers", optional = true }
